Update to nuget packages, Fix bug with odometer.

+++ TpmsCommon/Classes/BaseVehicle.cs
@@ -8,8 +8,8 @@ namespace TPMSIoTDemo.Common
     {
         public Guid Id { get; set; }
         public virtual List<BaseVehicleTire> Tires { get; set; }
-        public int OdometerInMiles { get; set; }
-        public virtual void Move(int Increment) { }
+        public double OdometerInMiles { get; set; }
+        public virtual void Move(double Increment) { }
         public string FactoryName { get; set; }
         public VehicleType VehicleType { get; set; }
         public string VehicleClass { get; set; }
@@ -18,6 +18,6 @@ namespace TPMSIoTDemo.Common
         public virtual void Stop() { }
         public virtual VehicleTireReading ReadTires() { return new VehicleTireReading(); }
         public virtual void ReplaceFlat() { }
-        public virtual void Slow(int Increment) { }
+        public virtual void Slow(double Increment) { }
     }
 }
